Update naming scheme for userIdent
We don't want a single user to have lock access, we want to limit it per request.

It makes more sense to just pass it a key which _could_ be a user ID, but more generally is just the ID of whatever we want to own the lock, be it a user or a request.