Fix behavior when a given userIdent tries to obtain a second lock on a record
Earlier we fixed Doctrine_Locking_Manager_Pessimistic to object_keys of the record's ID rather than the string 'id'. However, when trying to obtain a second lock, it was skipping the getLockOwner() method, and therefore using the string 'id'. This fixes that behavior so a given userIdent can obtain a lock multiple times on a record.